@@ -23,7 +23,7 @@
     "@vue/cli-plugin-eslint": "~4.4.0",
     "@vue/cli-service": "~4.4.0",
     "babel-eslint": "^10.1.0",
-    "electron": "^8.3.4",
+    "electron": "^9.0.5",
     "electron-rebuild": "^1.11.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
