@@ -25,31 +25,17 @@
     </form>
     <div class="max-w-xl p-10 m-auto">
       <ul>
-        <li v-for="reminder in reminders" :key="reminder.id" class="flex space-y-2 items-baseline hover:bg-gray-700 px-4 py-1 rounded-sm">
-          <clock-icon size="1x" class="text-blue-300"></clock-icon> 
-          <p class="text-md tracking-wide text-white ml-3 space-y-1">
-            <span class="mx-3" v-text="getDate(reminder.date)"></span>  
-            <span class="mx-3">{{ reminder.time }}</span> 
-            <span class="mx-3">{{reminder.message}}</span> 
-          </p>
-        <button class="rounded-sm bg-red-500 shadow-lg p-1 text-white ml-auto transition duration-300 ease-in-out transform hover:translate-x-1" type="button"
-          v-on:click="deleteReminder(reminder.id)">
-          <trash-2-icon size="1x"></trash-2-icon> 
-        </button>
+        <li v-for="reminder in reminders" :key="reminder.id">
+          <Reminder :reminder="reminder"></Reminder>
         </li>
       </ul>
     </div>
-    <HelloWorld/>
   </div>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld.vue'
-import { Trash2Icon, ClockIcon } from 'vue-feather-icons'
-// import {addJob} from './jobs'; 
-
-const notifier = require('node-notifier');
-
+import Reminder from './components/Reminder'
+import {addJob} from './jobs'; 
 
 export default {
   name: 'App',
@@ -62,9 +48,7 @@ export default {
     }
   },
   components: {
-    HelloWorld,
-    Trash2Icon,
-    ClockIcon 
+    Reminder,
   },
   methods: {
     saveReminder(e) {
@@ -79,23 +63,12 @@ export default {
       }
       this.reminders.push(newReminder);
       localStorage.reminders = JSON.stringify(this.reminders);
-      // addJob(newReminder);
-       notifier.notify({
-                title: "This is a notif",
-                message: "Hello, swati"
-            });
+      addJob(newReminder);
     },
-    deleteReminder(id) {
-      this.reminders = this.reminders.filter(item => item.id != id);
-      localStorage.reminders = JSON.stringify(this.reminders);
-    },
-    getDate(dateString) {
-      return new Date(dateString).toLocaleDateString('en-GB');
-    }
   },
   mounted() {
     this.reminders = JSON.parse(localStorage.reminders) || [];
-  },
+  }
 }
 </script>
 
