@@ -1,21 +1,28 @@
-// export default addJob;
 const cronJob = require('cron').CronJob;
-const notifier = require('node-notifier');
+const {ipcRenderer} = window.require('electron');
 
-const jobs = [];
+let jobs = [];
 
 function addJob(reminder) {
     const date = reminder.date.split('-');
     const time = reminder.time.split(':');
     date[1] = date[1] === '01' ? '0' : date[1] - 1;
-    jobs.push(
-        new cronJob( `00 ${time[1]} ${time[0]} ${date[2]} ${date[1]} *`, function() {
-            notifier.notify({
-                title: "This is a notification",
-                message: "Hello, swati"
-            })
-        }, null, true)
+
+    jobs.push({
+            id: reminder.id,
+            cb: new cronJob( `00 ${time[1]} ${time[0]} ${date[2]} ${date[1]} *`, function() {
+                ipcRenderer.send('notify', reminder.message);
+            }, null, true)
+        }
     );
 }
 
-export {addJob}; 
+function deleteJob(id) {
+    jobs.forEach(job => job.cb.stop());
+
+    jobs = jobs.filter(job => job.id != id);
+
+    jobs.forEach(job => job.cb.start());
+}
+
+export { addJob, deleteJob }; 
